backend/requirements.txt
langchain==0.1.20
langgraph==0.0.40
langchain-openai==0.1.7
openai==1.37.0
fastapi==0.104.0
uvicorn==0.24.0
websockets==11.0.3
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
